refactor: move thread runtime run cancellation
